создаёт класс game-view и переносит туда дублирующуюся логику
